moved around files, and tried to get custom renderers to work
