+++ Mobile/SecurityInAMobile/SecurityInAMobile/SecurityInAMobile.Android/SecurityInAMobile.Android.csproj
@@ -55,7 +55,6 @@
     <PackageReference Include="Xamarin.Android.Support.v7.MediaRouter" Version="25.4.0.2" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="CustomSwitchRenderer.cs" />
     <Compile Include="MainActivity.cs" />
     <Compile Include="Resources\Resource.Designer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
