+++ Mobile/SecurityInAMobile/SecurityInAMobile/SecurityInAMobile/SecurityInAMobile.projitems
@@ -16,6 +16,8 @@
     <Compile Include="$(MSBuildThisFileDirectory)App.xaml.cs">
       <DependentUpon>App.xaml</DependentUpon>
     </Compile>
+    <Compile Include="$(MSBuildThisFileDirectory)CustomPickerRenderer.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)CustomSwitchRenderer.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)MainPage.xaml.cs">
       <DependentUpon>MainPage.xaml</DependentUpon>
     </Compile>
